Moved the tests into exsiting files
Rename some of the classes to match their environment
Added binding=False test
