+++ tests/run/staticmethod.pyx
@@ -1,39 +1,17 @@
-import sys
-
-__doc__ = u"""
->>> class4.plus1(1)
-2
->>> class4().plus1(1)
-2
->>> class4.bplus1(1)
-2
->>> class4().bplus1(1)
-2
-"""
-# for class3, follow Python behaviour tested in method_assignment.py
-# (This class hasn't been moved since it isn't pure Python)
-if sys.version_info[0] == 2:
-    __doc__ += """>>> class3.plus1(1) #doctest: +IGNORE_EXCEPTION_DETAIL
-Traceback (most recent call last):
-    ...
-TypeError: unbound
-"""
-else:
-    __doc__ += """>>> class3.plus1(1)
-2
-"""
-
 cimport cython
 
-def f_plus(a):
-    return a + 1
 
-# The tests formerly in class1 and class2 have been moved to "method_assignment.py"
-
-cdef class class3:
-    plus1 = f_plus
-
-class class4:
+class class1:
+    u"""
+    >>> class1.plus1(1)
+    2
+    >>> class1().plus1(1)
+    2
+    >>> class1.bplus1(1)
+    2
+    >>> class1().bplus1(1)
+    2
+    """
     @staticmethod
     def plus1(a):
         return a + 1
@@ -53,14 +31,14 @@ def nested_class():
     >>> obj.plus1(1)
     2
     """
-    class class5(object):
+    class class2(object):
         def __new__(cls): # implicit staticmethod
             return object.__new__(cls)
 
         @staticmethod
         def plus1(a):
             return a + 1
-    return class5
+    return class2
 
 
 cdef class BaseClass(object):
